# feat: improve `make_date` performance