@alamb 2024-02-02
I also ran the benchmarks:
```shell
git checkout a6ef1bec480872f15f83628a7fb8c9bb2722cd49 # merge-base
```

results are basically the same as yours:

```
alamb@aal-dev:~/arrow-datafusion3$ cargo bench --bench make_date
   Compiling datafusion-physical-expr v35.0.0 (/home/alamb/arrow-datafusion3/datafusion/physical-expr)
    Finished bench [optimized] target(s) in 2m 46s
     Running benches/make_date.rs (target/release/deps/make_date-4622d805a0644623)
Gnuplot not found, using plotters backend
make_date_col_col_col_1000
                        time:   [10.392 µs 10.399 µs 10.410 µs]
                        change: [-57.725% -57.673% -57.624%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) high mild
  4 (4.00%) high severe

make_date_scalar_col_col_1000
                        time:   [11.036 µs 11.048 µs 11.060 µs]
                        change: [-48.876% -48.746% -48.575%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high severe

make_date_scalar_scalar_col_1000
                        time:   [11.386 µs 11.400 µs 11.415 µs]
                        change: [-37.743% -37.655% -37.562%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

make_date_scalar_scalar_scalar
                        time:   [1.2737 µs 1.2759 µs 1.2783 µs]
                        change: [+2.3366% +2.7260% +3.2257%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 17 outliers among 100 measurements (17.00%)
  2 (2.00%) low severe
  2 (2.00%) low mild
  4 (4.00%) high mild
  9 (9.00%) high severe

```

🚀