@r3stl355 2024-02-02
Yes, I had some regression on all scalar test (`make_date_scalar_scalar_scalar`) but I managed to improve it with if/else - it's minimal now compared to improvements you get when at least one is an array