@Omega359 2024-02-03
The 
```let array_size = if is_scalar { 1 } else { len.unwrap() };
```
variable could now be removed and the for loop could just use len.unwrap() I think.