@r3stl355 2024-02-03
> The
> 
> `let array_size = if is_scalar { 1 } else { len.unwrap() };`
> 
> variable could now be removed and the for loop could just use len.unwrap() I think.

Sure, I can do that but `array_size` is also used in the 3 calls to `to_primitive_array`. Alternatively, I could just move the variable into `else` as `let array_size = len.unwrap();`. What would be your preference @Omega359 ?