@Omega359 2024-02-03
> Sure, I can do that but `array_size` is also used in the 3 calls to `to_primitive_array`. Alternatively, I could just move the variable into `else` as `let array_size = len.unwrap();`. What would be your preference @Omega359 ?

Hmm, I think I missed that - the downside of looking at code outside of an IDE without proper highlighting. Please ignore my comment then :) 

I have to see if there is a way to look at and comment on these pull requests easily in rust rover. Hopefully they've brought that functionality over from IntelliJ IDEA.