@r3stl355 2024-02-04
I've not tried RustRover, mainly work with VSCode. I don't have JetBrains licenses and though RR is free for now, I don't want to put time into learning something I won't be able to use once it becomes paid-for 🤷