@alamb 2024-02-04
> I have to see if there is a way to look at and comment on these pull requests easily in rust rover. Hopefully they've brought that functionality over from IntelliJ IDEA.

I believe this feature is present (though I normally just use the github UI or else check out the PR directly via `gh pr checkout <url>`